Improves UI and adds logout functionality
Enhances the user interface of income, expense, and installment pages by adding cancel buttons and transitions.

Adds a logout button and moves the navigation bar to a separate component for better code organization and reusability.

Improves user authentication by redirecting to the login page if the user is not authenticated.